fix processVideo in cyclic depend
