+++ lib/v2/bloomberg/react-renderer.js
@@ -1,6 +1,39 @@
 const art = require('art-template');
 const path = require('path');
-const { processVideo } = require('./utils');
+const got = require('@/utils/got');
+
+const headers = {
+    accept: 'application/json',
+    'cache-control': 'no-cache',
+    referer: 'https://www.bloomberg.com',
+};
+
+const processVideo = async (bmmrId, summary) => {
+    const api = `https://www.bloomberg.com/multimedia/api/embed?id=${bmmrId}`;
+    const res = await got(api, { headers });
+
+    // Blocked by PX3, return the default
+    const redirectUrls = res.redirectUrls.map(String);
+    if (redirectUrls.some((r) => new URL(r).pathname === '/tosv2.html')) {
+        return {
+            stream: '',
+            mp4: '',
+            coverUrl: '',
+            caption: summary,
+        };
+    }
+
+    if (res.data) {
+        const video_json = res.data;
+        return {
+            stream: video_json.streams ? video_json.streams[0]?.url : '',
+            mp4: video_json.downloadURLs ? video_json.downloadURLs['600'] : '',
+            coverUrl: video_json.thumbnail?.baseUrl ?? '',
+            caption: video_json.description || video_json.title || summary,
+        };
+    }
+    return {};
+};
 
 const nodeRenderers = {
     paragraph: async (node, nextNode) => `<p>${await nextNode(node.content)}</p>`,
@@ -128,9 +161,9 @@ const nodeRenderers = {
         }
         if (t === 'video') {
             const h = node.data;
-            const id = h.attachment && h.attachment.id;
-            const v = await processVideo(id);
-            return v;
+            const id = h.attachment?.id;
+            const desc = await processVideo(id, h.attachment?.title);
+            return art(path.join(__dirname, 'templates/video_media.art'), desc);
         }
         if (t === 'audio' && node.data.attachment) {
             const B = node.data.attachment;
@@ -187,4 +220,6 @@ const documentToHtmlString = async (document) => {
 
 module.exports = {
     documentToHtmlString,
+    processVideo,
+    headers,
 };
