+++ lib/v2/bloomberg/utils.js
@@ -6,7 +6,7 @@ const { parseDate } = require('@/utils/parse-date');
 const got = require('@/utils/got');
 const { art } = require('@/utils/render');
 
-const { documentToHtmlString } = require('./react-renderer');
+const { documentToHtmlString, processVideo, headers } = require('./react-renderer');
 
 const rootUrl = 'https://www.bloomberg.com/feeds';
 const sel = 'script[data-component-props="ArticleBody"], script[data-component-props="FeatureBody"]';
@@ -39,11 +39,6 @@ const apiEndpoints = {
         url: 'https://www.bloomberg.com/features/',
         sel: 'script#__SSR_DATA__',
     },
-};
-const headers = {
-    accept: 'application/json',
-    'cache-control': 'no-cache',
-    referer: 'https://www.bloomberg.com',
 };
 
 const pageTypeRegex1 = /\/(?<page>[\w-]*?)\/(?<link>\d{4}-\d{2}-\d{2}\/.*)/;
@@ -83,7 +78,7 @@ const parseArticle = (item, ctx) =>
             .map((a) => a && a.groups)[0];
         if (group) {
             const { page, link } = group;
-            if (apiEndpoints[page]) {
+            if (apiEndpoints[page] && item.link === 'https://www.bloomberg.com/news/articles/2023-07-25/cambio-de-marca-twitter-eliminara-una-fortuna-en-valor-de-marca') {
                 const api = { ...apiEndpoints[page] };
                 let res;
 
@@ -337,33 +332,6 @@ const processLedeMedia = async (story_json) => {
     }
 };
 
-const processVideo = async (bmmrId, summary) => {
-    const api = `https://www.bloomberg.com/multimedia/api/embed?id=${bmmrId}`;
-    const res = await got(api, { headers });
-
-    // Blocked by PX3, return the default
-    const redirectUrls = res.redirectUrls.map(String);
-    if (redirectUrls.some((r) => new URL(r).pathname === '/tosv2.html')) {
-        return {
-            stream: '',
-            mp4: '',
-            coverUrl: '',
-            caption: summary,
-        };
-    }
-
-    if (res.data) {
-        const video_json = res.data;
-        return {
-            stream: video_json.streams ? video_json.streams[0]?.url : '',
-            mp4: video_json.downloadURLs ? video_json.downloadURLs['600'] : '',
-            coverUrl: video_json.thumbnail?.baseUrl ?? '',
-            caption: video_json.description || video_json.title || summary,
-        };
-    }
-    return {};
-};
-
 const processBody = async (body_html, story_json) => {
     const removeSel = ['meta', 'script', '*[class$="-footnotes"]', '*[class$="for-you"]', '*[class$="-newsletter"]', '*[class$="page-ad"]', '*[class$="-recirc"]', '*[data-ad-placeholder="Advertisement"]'];
 
@@ -441,5 +409,4 @@ module.exports = {
     asyncPoolAll,
     parseNewsList,
     parseArticle,
-    processVideo,
 };
